user-manager/requirements.txt
flask==2.2.5
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
grpcio==1.60.0
grpcio-tools==1.60.0
apscheduler==3.10.4
prometheus-client
